+++ README.md
@@ -1,11 +1,11 @@
-Nerdbank.Algorithms
-===================
+# Nerdbank.Algorithms
 
 This project contains a collection of popular algorithms I've found useful in some fun projects.
 
 [![NuGet package](https://img.shields.io/nuget/v/Nerdbank.Algorithms.svg)](https://nuget.org/packages/Nerdbank.Algorithms)
 [![Build status](https://ci.appveyor.com/api/projects/status/rm4h7dql9t34xuvw/branch/master?svg=true)](https://ci.appveyor.com/project/AArnott/nerdbank-algorithms/branch/master)
 
-## Algorithms included
+## Constraints solution engine
 
- * [Brand and bound](https://en.wikipedia.org/wiki/Branch_and_bound)
+The `Nerdbank.Algorithms.NodeConstraintSelection.SolutionBuilder<T>` class will help to find the one (or many) solutions
+to any problem that you can express in terms of nodes and constraints.
