+++ src/NerdBank.Algorithms.sln
@@ -2,9 +2,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29622.13
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NerdBank.Algorithms", "NerdBank.Algorithms\NerdBank.Algorithms.csproj", "{7D0A3E71-C7A5-4465-BCA1-0B9533936D29}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nerdbank.Algorithms", "Nerdbank.Algorithms\Nerdbank.Algorithms.csproj", "{7D0A3E71-C7A5-4465-BCA1-0B9533936D29}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NerdBank.AlgorithmsTest", "NerdBank.AlgorithmsTest\NerdBank.AlgorithmsTest.csproj", "{B4474D3C-8D00-46CE-9616-A7663967E65C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nerdbank.AlgorithmsTest", "Nerdbank.AlgorithmsTest\Nerdbank.AlgorithmsTest.csproj", "{B4474D3C-8D00-46CE-9616-A7663967E65C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D5505154-F37C-47AD-A881-0F985B28D2E0}"
 	ProjectSection(SolutionItems) = preProject
