+++ src/Nerdbank.Algorithms/BannedSymbols.txt
