+++ src/Nerdbank.Algorithms/Nerdbank.Algorithms.csproj
