+++ src/Nerdbank.Algorithms/NodeConstraintSelection/BadConstraintException`1.cs
@@ -1,52 +1,52 @@
 ﻿// Copyright (c) Andrew Arnott. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace NerdBank.Algorithms.NodeConstraintSelection
+namespace Nerdbank.Algorithms.NodeConstraintSelection
 {
 	using System;
 	using System.Runtime.Serialization;
 
 	/// <summary>
-	/// An exception that identifies when a combination of constraints cannot be simultaneously satisfied.
+	/// An exception thrown when some <see cref="IConstraint{TNodeState}"/> misbehaves.
 	/// </summary>
 	/// <typeparam name="TNodeState">The type of value that a node may be set to.</typeparam>
 	[Serializable]
 #pragma warning disable CA1032 // Implement standard exception constructors
-	public class BrokenConstraintException<TNodeState> : Exception
+	public class BadConstraintException<TNodeState> : Exception
 		where TNodeState : struct
 #pragma warning restore CA1032 // Implement standard exception constructors
 	{
-		/// <inheritdoc cref="BrokenConstraintException{TNodeState}(IConstraint{TNodeState}, string, Exception)"/>
-		public BrokenConstraintException(IConstraint<TNodeState> constraint)
+		/// <inheritdoc cref="BadConstraintException{TNodeState}(IConstraint{TNodeState}, string, Exception)"/>
+		public BadConstraintException(IConstraint<TNodeState> constraint)
 		{
 			this.Constraint = constraint ?? throw new ArgumentNullException(nameof(constraint));
 		}
 
-		/// <inheritdoc cref="BrokenConstraintException{TNodeState}(IConstraint{TNodeState}, string, Exception)"/>
-		public BrokenConstraintException(IConstraint<TNodeState> constraint, string message)
+		/// <inheritdoc cref="BadConstraintException{TNodeState}(IConstraint{TNodeState}, string, Exception)"/>
+		public BadConstraintException(IConstraint<TNodeState> constraint, string message)
 			: base(message)
 		{
 			this.Constraint = constraint ?? throw new ArgumentNullException(nameof(constraint));
 		}
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="BrokenConstraintException{TNodeState}"/> class.
+		/// Initializes a new instance of the <see cref="BadConstraintException{TNodeState}"/> class.
 		/// </summary>
 		/// <param name="constraint">The constraint that misbehaved.</param>
 		/// <param name="message">A message about how the constraint misbehaved.</param>
 		/// <param name="inner">An inner exception.</param>
-		public BrokenConstraintException(IConstraint<TNodeState> constraint, string message, Exception inner)
+		public BadConstraintException(IConstraint<TNodeState> constraint, string message, Exception inner)
 			: base(message, inner)
 		{
 			this.Constraint = constraint ?? throw new ArgumentNullException(nameof(constraint));
 		}
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="BrokenConstraintException{TNodeState}"/> class.
+		/// Initializes a new instance of the <see cref="BadConstraintException{TNodeState}"/> class.
 		/// </summary>
 		/// <param name="info">Serialization info.</param>
 		/// <param name="context">Serialization context.</param>
-		protected BrokenConstraintException(SerializationInfo info, StreamingContext context)
+		protected BadConstraintException(SerializationInfo info, StreamingContext context)
 			: base(info, context)
 		{
 			this.Constraint = (IConstraint<TNodeState>)info.GetValue(nameof(this.Constraint), typeof(IConstraint<TNodeState>));
