+++ src/Nerdbank.Algorithms/NodeConstraintSelection/Strings.resx
