+++ src/Nerdbank.Algorithms/Properties/AssemblyInfo.cs
