+++ src/Nerdbank.AlgorithmsTest/NodeConstraintSelection/DummyNode.cs
