+++ src/Nerdbank.AlgorithmsTest/TestBase.cs
