+++ src/Nerdbank.AlgorithmsTest/app.config
